BUGFIX: Clickmenu did not appear or did appear on the wrong position in the filelist module using Mozilla. (I only tested that bug with Mozilla / Linux).
REASON:
In the JavaScript function setLayerObj I discovered that when the menu does not appear, the tempLayerObjCss.top and tempLayerObjCss.left properties contained nothing. It turned out that explicitly setting their values to "10px" instead of "10" made the menu appear correctly.

SOLUTION:
I haven't found yet another place in the core which might be affected by this bug, so I only changed the lines in the JavaScript function. Maybe someoneelse knows about other places we have to look at?